src/requirements.txt
braintree==3.25.0
cffi==1.5.2
cryptography==1.3.1
Django==1.9.3
django-cors-headers==1.1.0
django-crispy-forms==1.6.0
django-filter==0.13.0
django-registration-redux==1.4
djangorestframework==3.3.3
djangorestframework-jwt==1.7.2
idna==2.1
ipaddress==1.0.16
Markdown==2.6.6
ndg-httpsclient==0.4.0
Pillow==3.1.1
pyasn1==0.1.9
pycparser==2.14
PyJWT==1.4.0
pyOpenSSL==16.0.0
requests==2.9.1
six==1.10.0
